+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - documenteditor/config/editorconfig/Customization#mobile
 - documenteditor/config/editorconfig/Customization#pointerMode
 - documenteditor/config/editorconfig/Customization#slidePlayerBackground
+- documenteditor/config/editorconfig/Customization#wordHeadingsColor
 - deprecated documenteditor/config/editorconfig/Customization#mobileForceView
 - support hwp, hwpx, pages, numbers, key formats
 - sq-AL empty file templates